+++ blocks/course_overview/save.php
@@ -34,6 +34,7 @@ $activetab = required_param('tab', PARAM_ALPHA);
 
 if ($activetab == 'courses') {
     block_course_overview_update_myorder($sortorder);
+    block_course_overview_update_sortorder(0);
 } else if ($activetab == "favourites") {
     block_course_overview_update_favourites($sortorder);
 }